api: refactor activities by creating activities.util.ts
